+++ src/Administration/ImportExportRedirectsController.php
@@ -103,9 +103,9 @@ class ImportExportRedirectsController extends AbstractController
         /**
          * @var Redirect $redirect
          */
-        fputcsv($fileStream, array('id', 'source_url', "target_url", "http_code", "enabled", "ignore_query_params"), ';'); // Writing the CSV Headline
+        fputcsv($fileStream, array('id', 'source_url', "target_url", "http_code", "enabled", "query_params_handling"), ';'); // Writing the CSV Headline
         foreach ($redirects as $redirect) { // Writing each Redirect into the file
-            fputcsv($fileStream, array($redirect->getId(), $redirect->getSourceURL(), $redirect->getTargetURL(), $redirect->getHttpCode(), $redirect->isEnabled() ? 1 : 0, $redirect->isIgnoreQueryParams() ? 1 : 0), ';');
+            fputcsv($fileStream, array($redirect->getId(), $redirect->getSourceURL(), $redirect->getTargetURL(), $redirect->getHttpCode(), $redirect->isEnabled() ? 1 : 0, $redirect->getQueryParamsHandling()), ';');
         }
 
         //Closing the File
@@ -224,12 +224,12 @@ class ImportExportRedirectsController extends AbstractController
 
         $valid = false;
         if (count($title) === 5) { // Backward compatibility for exports before v1.2.0
-            if ($title[0] === "id" || $title[1] === "source_url" || $title[2] === "target_url" || $title[3] === "http_code" || $title[4] === "enabled") {
+            if ($title[0] === "id" && $title[1] === "source_url" && $title[2] === "target_url" && $title[3] === "http_code" && $title[4] === "enabled") {
                 $valid = true;
             }
         }
         if (count($title) === 6) {
-            if ($title[0] === "id" || $title[1] === "source_url" || $title[2] === "target_url" || $title[3] === "http_code" || $title[4] === "enabled" || $title[5] === "ignore_query_params") {
+            if ($title[0] === "id" && $title[1] === "source_url" && $title[2] === "target_url" && $title[3] === "http_code" && $title[4] === "enabled" && ($title[5] === "query_params_handling" || $title[5] === "ignore_query_params")) { // ignore_query_params: Backward compatibility for exports before v2.1.0
                 $valid = true;
             }
         }
@@ -250,13 +250,13 @@ class ImportExportRedirectsController extends AbstractController
             $httpCode = intval($line[3]);
 
             //Checking if this line has invalid data
-            if ($sourceURL === "" || ($httpCode != 301 && $httpCode != 302) || ($line[4] !== "1" && $line[4] !== "0") || (count($line) === 6 && $line[5] !== "1" && $line[5] !== "0")) {
+            if ($sourceURL === "" || ($httpCode != 301 && $httpCode != 302) || ($line[4] !== "1" && $line[4] !== "0") || (count($line) === 6 && $line[5] !== "2" && $line[5] !== "1" && $line[5] !== "0")) {
                 $error++;
                 continue;
             }
 
             $enabled = boolval($line[4]);
-            $ignoreQueryParams = count($line) === 6 && boolval($line[5]);
+            $queryParamsHandling = count($line) === 6 ? intval($line[5]) : 0;
 
             //Search either for the id and the sourceURL in the Database, or if the id is empty, only for the sourceURL
             $criteria = new Criteria();
@@ -285,7 +285,7 @@ class ImportExportRedirectsController extends AbstractController
                         'targetURL' => $targetURL,
                         'httpCode' => $httpCode,
                         'enabled' => $enabled,
-                        'ignoreQueryParams' => $ignoreQueryParams
+                        'queryParamsHandling' => $queryParamsHandling
                     ]], $context);
                     $count++;
                 } else if (strcasecmp($redirect->getSourceURL(), $sourceURL) == 0 && $override) { //The SourceURLs match and should be updated, updating it
@@ -295,7 +295,7 @@ class ImportExportRedirectsController extends AbstractController
                         'targetURL' => $targetURL,
                         'httpCode' => $httpCode,
                         'enabled' => $enabled,
-                        'ignoreQueryParams' => $ignoreQueryParams
+                        'queryParamsHandling' => $queryParamsHandling
                     ]], $context);
                     $count++;
                 } else { //The Redirect should not be overridden, skipping it
@@ -310,7 +310,7 @@ class ImportExportRedirectsController extends AbstractController
                         'targetURL' => $targetURL,
                         'httpCode' => $httpCode,
                         'enabled' => $enabled,
-                        'ignoreQueryParams' => $ignoreQueryParams
+                        'queryParamsHandling' => $queryParamsHandling
                     ]], $context);
                 else //... with a new ID
                     $this->redirectRepository->create([[
@@ -318,7 +318,7 @@ class ImportExportRedirectsController extends AbstractController
                         'targetURL' => $targetURL,
                         'httpCode' => $httpCode,
                         'enabled' => $enabled,
-                        'ignoreQueryParams' => $ignoreQueryParams
+                        'queryParamsHandling' => $queryParamsHandling
                     ]], $context);
                 $count++;
             }